Move EmbraceCustomPathException back to original package
