+++ embrace-android-instrumentation-okhttp/src/main/kotlin/io/embrace/android/embracesdk/okhttp3/EmbraceCustomPathException.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.embracesdk.internal.instrumentation.okhttp
+package io.embrace.android.embracesdk.okhttp3
 
 import java.io.IOException
